tools/skill_generator.py
import yaml
from rich.console import Console

console = Console()

def generate_skill():
    """
    模拟合并 prompts 和语料生成最终 SKILL.md 的逻辑
    """
    console.print("[bold green]正在初始化震哥逻辑引擎...[/bold green]")
    console.print("[yellow]正在注入‘这你受得了吗’高亢情绪模块...[/yellow]")
    
    # 模拟从 SKILL.md 读取元数据
    with open('SKILL.md', 'r', encoding='utf-8') as f:
        content = f.read()
        
    console.print("\n[bold cyan]震哥.skill 生成报告：[/bold cyan]")
    console.print("- 版本: 1.2 (硬核版)")
    console.print("- 核心逻辑: 性能/效率/价值三维分析")
    console.print("- 灵魂词汇: [bold red]这你受得了吗？[/bold red]")
    
    console.print("\n[green]Skill 构建成功。这性能，这效率，没毛病吧？是吧？[/green]")

if __name__ == "__main__":
    generate_skill()
