requirements.txt
jinja2>=3.0.0
pyyaml>=6.0
rich>=12.0.0
